Remove TS diagnostic suppression
